feat: add api update books

+++ core/helpers.php
@@ -200,11 +200,12 @@ if (! function_exists('removeFile')) {
     /**
      * Remove file based on filename.
      *
-     * @param string $filename
+     * @param string|null $filename
      * @return void
      */
-    function removeFile(string $filename): void
+    function removeFile(string|null $filename): void
     {
+        if (! $filename) return;
         if (file_exists($filename)) {
             unlink($filename);
         }
